+++ src/add-ons/media/plugins/musepack/MusePackDecoder.h
@@ -20,7 +20,7 @@ class MusePackDecoder : public Decoder {
 		void GetCodecInfo(media_codec_info *info);
 
 		status_t Setup(media_format *ioEncodedFormat,
-					const void *infoBuffer, int32 infoSize);
+					const void *infoBuffer, size_t infoSize);
 
 		status_t NegotiateOutputFormat(media_format *ioDecodedFormat);
