+++ src/add-ons/media/plugins/vorbis/vorbisCodecPlugin.h
@@ -13,7 +13,7 @@ public:
 	
 	void		GetCodecInfo(media_codec_info *info);
 	status_t	Setup(media_format *inputFormat,
-					  const void *infoBuffer, int32 infoSize);
+					  const void *infoBuffer, size_t infoSize);
 
 	status_t	NegotiateOutputFormat(media_format *ioDecodedFormat);
